+++ src/BDSDetectorConstruction.cc
@@ -219,6 +219,13 @@ void BDSDetectorConstruction::BuildBeamlines()
       exit(1);
     }
 
+  // print warning if beamline is approximately circular but flag isnt specfied
+  if (not circular and mainBeamline.massWorld->ElementAnglesSumToCircle())
+    {
+      G4cerr << __METHOD_NAME__ << "WARNING: Total sum of all element angles is approximately 2*pi"
+             << " but the circular option was not specified, this simulation may run indefinitely" << G4endl;
+    }
+
   // register the beamline in the holder class for the full model
   acceleratorModel->RegisterBeamlineSetMain(mainBeamline);